資料庫調整

+++ Tokiku.Entity/ViewTables/AluminumExtrusionOrderMaterialValuationEntity.cs
@@ -13,7 +13,7 @@ namespace Tokiku.Entity
         //"*kg單價*"
         public Nullable<int> UnitPrice { get; set; }
         //"重量"
-        public Nullable<float> Weight { get; set; }
+        public Nullable<double> Weight { get; set; }
         //"預估總價"
         public Nullable<int> TotalPrice { get; set; }
     }
